Reorganize `process_events` module
